Remove admin reducers and actions, and move the functions under hearingEditor
