+++ src/reducers/index.js
@@ -7,7 +7,6 @@ import hearing from './hearing';
 import sectionComments from './sectionComments';
 import labels from './labels';
 import hearingEditor from './hearingEditor';
-import admin from './admin';
 
 /**
 A reducer that stores the type -- and _only_ the type --
@@ -29,6 +28,5 @@ export default combineReducers({
   router,
   sectionComments,
   user,
-  labels,
-  admin
+  labels
 });
